training/requirements.txt
numpy==1.24.4
pandas==2.0.3
torch==2.4.1
torchvision==0.19.1
transformers==4.34.0
sentence-transformers==2.2.2
onnx==1.17.0
onnxruntime==1.19.2
faiss-cpu==1.8.0.post1
scikit-learn==1.3.2
tqdm==4.67.1